# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/api/v1/propertyInfo": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "物业信息"
                ],
                "summary": "根据条件获取物业信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "行业ID",
                        "name": "industry_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "店铺类型ID",
                        "name": "store_type_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "省代码",
                        "name": "province_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "城市代码",
                        "name": "city_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "区代码",
                        "name": "district_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "街道代码",
                        "name": "street_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "面积范围ID",
                        "name": "area_type_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "租金范围ID",
                        "name": "rent_type_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "最小面积",
                        "name": "min_area",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "最大面积",
                        "name": "max_area",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "最小租金",
                        "name": "min_rent",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "最大租金",
                        "name": "max_rent",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "模型类型 0-转让 ｜ 1-出售 ｜ 3-出租 | 4-求租 ｜ 5-求购",
                        "name": "model_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "业务类型 0-商铺 ｜ 1-写字楼 ｜ 2-厂房仓库",
                        "name": "bus_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序 area-面积 ｜ rent-租金 ｜ created_at-发布时间（默认）",
                        "name": "sort_condition",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "页大小 （默认30）",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "跳转页码",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/propertyInfo/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "物业信息"
                ],
                "summary": "查看物业详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "物业信息ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "查看用户信息",
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "修改用户信息",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.EditUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/user/edituserpass": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "修改用户密码",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.EditUserPass"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/user/findStore": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "修改我要找铺信息",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UserFindStoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "我要找铺",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UserFindStoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/user/propertyInfo": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "查看用户已发布物业信息",
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/user/propertyInfo/{id}/picture": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "上传图集图片（单张）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "物业信息ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "图片",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/user/propertyInfo/{pro_id}/picture/{pri_id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "删除图片（单张）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "物业信息ID",
                        "name": "pro_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "图片ID",
                        "name": "pri_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.EditUser": {
            "type": "object",
            "properties": {
                "QQ": {
                    "description": "QQ",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "landlinenumber": {
                    "description": "座机号码",
                    "type": "string"
                },
                "nickname": {
                    "description": "姓名",
                    "type": "string"
                },
                "sex": {
                    "description": "性别 0:未知 ｜ 1:男 ｜ 2:女 （空或其他默认未知）",
                    "type": "string"
                },
                "telephone": {
                    "description": "手机号码",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "entity.EditUserPass": {
            "type": "object",
            "properties": {
                "new_pass": {
                    "description": "新密码",
                    "type": "string"
                },
                "old_pass": {
                    "description": "旧密码",
                    "type": "string"
                }
            }
        },
...
        "entity.ResponseData": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "返回数据",
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                },
                "status": {
                    "description": "成功失败标志；true：成功 、false：失败",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}